Initial commit for ESP8266 Power Controller
Sketch to configure a ESP8266 board (tested on Lolin D1 Mini) to send
433MHz RF commands in response to web requests.  Uses the
ESP8266WebServer, ESP8266mDNS, and RCSwitch libraries.

The homepage and list of switch codes are currently hardcoded, as are
the WiFi credentials and hostname ("power".)  You can sniff the codes
using the RCSwitch ReceiveDemo_Advanced script.

Portions based on the ESP8266 PathArgServer demo and RCSwitch Webserver
demo.